requirements.txt
python-dotenv==0.21.0
discord==2.0.0
jellyfish==0.9.0
pydantic==1.10.2
SQLAlchemy==1.4.41